requirements.txt
aiofiles==0.7.0
aiosqlite==0.17.0
alembic==1.6.5
asyncpg==0.23.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
databases==0.4.1
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
fastapi-users==6.1.0
h11==0.12.0
idna==3.2
Jinja2==3.0.1
makefun==1.11.3
Mako==1.1.4
MarkupSafe==2.0.1
ormar==0.10.9
passlib==1.7.4
psycopg2==2.8.6
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-multipart==0.0.5
six==1.16.0
SQLAlchemy==1.3.23
starlette==0.14.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
wrapt==1.12.1
